rules_api/api/requirements.txt
antlr4-python2-runtime
cabby
flask
kafka-python
stix2
stix2-validator